+++ sdk/scvmm/azure-resourcemanager-scvmm/src/main/java/com/azure/resourcemanager/scvmm/models/HardwareProfileUpdate.java
@@ -5,51 +5,56 @@
 package com.azure.resourcemanager.scvmm.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
-/** Defines the resource properties. */
+/**
+ * Defines the resource update properties.
+ */
 @Fluent
-public final class HardwareProfileUpdate {
+public final class HardwareProfileUpdate implements JsonSerializable<HardwareProfileUpdate> {
     /*
      * MemoryMB is the size of a virtual machine's memory, in MB.
      */
-    @JsonProperty(value = "memoryMB")
     private Integer memoryMB;
 
     /*
      * Gets or sets the number of vCPUs for the vm.
      */
-    @JsonProperty(value = "cpuCount")
     private Integer cpuCount;
 
     /*
-     * Gets or sets a value indicating whether to enable processor
-     * compatibility mode for live migration of VMs.
+     * Gets or sets a value indicating whether to enable processor compatibility mode for live migration of VMs.
      */
-    @JsonProperty(value = "limitCpuForMigration")
     private LimitCpuForMigration limitCpuForMigration;
 
     /*
      * Gets or sets a value indicating whether to enable dynamic memory or not.
      */
-    @JsonProperty(value = "dynamicMemoryEnabled")
     private DynamicMemoryEnabled dynamicMemoryEnabled;
 
     /*
      * Gets or sets the max dynamic memory for the vm.
      */
-    @JsonProperty(value = "dynamicMemoryMaxMB")
     private Integer dynamicMemoryMaxMB;
 
     /*
      * Gets or sets the min dynamic memory for the vm.
      */
-    @JsonProperty(value = "dynamicMemoryMinMB")
     private Integer dynamicMemoryMinMB;
 
     /**
+     * Creates an instance of HardwareProfileUpdate class.
+     */
+    public HardwareProfileUpdate() {
+    }
+
+    /**
      * Get the memoryMB property: MemoryMB is the size of a virtual machine's memory, in MB.
-     *
+     * 
      * @return the memoryMB value.
      */
     public Integer memoryMB() {
@@ -58,7 +63,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Set the memoryMB property: MemoryMB is the size of a virtual machine's memory, in MB.
-     *
+     * 
      * @param memoryMB the memoryMB value to set.
      * @return the HardwareProfileUpdate object itself.
      */
@@ -69,7 +74,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Get the cpuCount property: Gets or sets the number of vCPUs for the vm.
-     *
+     * 
      * @return the cpuCount value.
      */
     public Integer cpuCount() {
@@ -78,7 +83,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Set the cpuCount property: Gets or sets the number of vCPUs for the vm.
-     *
+     * 
      * @param cpuCount the cpuCount value to set.
      * @return the HardwareProfileUpdate object itself.
      */
@@ -90,7 +95,7 @@ public final class HardwareProfileUpdate {
     /**
      * Get the limitCpuForMigration property: Gets or sets a value indicating whether to enable processor compatibility
      * mode for live migration of VMs.
-     *
+     * 
      * @return the limitCpuForMigration value.
      */
     public LimitCpuForMigration limitCpuForMigration() {
@@ -100,7 +105,7 @@ public final class HardwareProfileUpdate {
     /**
      * Set the limitCpuForMigration property: Gets or sets a value indicating whether to enable processor compatibility
      * mode for live migration of VMs.
-     *
+     * 
      * @param limitCpuForMigration the limitCpuForMigration value to set.
      * @return the HardwareProfileUpdate object itself.
      */
@@ -111,7 +116,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Get the dynamicMemoryEnabled property: Gets or sets a value indicating whether to enable dynamic memory or not.
-     *
+     * 
      * @return the dynamicMemoryEnabled value.
      */
     public DynamicMemoryEnabled dynamicMemoryEnabled() {
@@ -120,7 +125,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Set the dynamicMemoryEnabled property: Gets or sets a value indicating whether to enable dynamic memory or not.
-     *
+     * 
      * @param dynamicMemoryEnabled the dynamicMemoryEnabled value to set.
      * @return the HardwareProfileUpdate object itself.
      */
@@ -131,7 +136,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Get the dynamicMemoryMaxMB property: Gets or sets the max dynamic memory for the vm.
-     *
+     * 
      * @return the dynamicMemoryMaxMB value.
      */
     public Integer dynamicMemoryMaxMB() {
@@ -140,7 +145,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Set the dynamicMemoryMaxMB property: Gets or sets the max dynamic memory for the vm.
-     *
+     * 
      * @param dynamicMemoryMaxMB the dynamicMemoryMaxMB value to set.
      * @return the HardwareProfileUpdate object itself.
      */
@@ -151,7 +156,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Get the dynamicMemoryMinMB property: Gets or sets the min dynamic memory for the vm.
-     *
+     * 
      * @return the dynamicMemoryMinMB value.
      */
     public Integer dynamicMemoryMinMB() {
@@ -160,7 +165,7 @@ public final class HardwareProfileUpdate {
 
     /**
      * Set the dynamicMemoryMinMB property: Gets or sets the min dynamic memory for the vm.
-     *
+     * 
      * @param dynamicMemoryMinMB the dynamicMemoryMinMB value to set.
      * @return the HardwareProfileUpdate object itself.
      */
@@ -171,9 +176,64 @@ public final class HardwareProfileUpdate {
 
     /**
      * Validates the instance.
-     *
+     * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeNumberField("memoryMB", this.memoryMB);
+        jsonWriter.writeNumberField("cpuCount", this.cpuCount);
+        jsonWriter.writeStringField("limitCpuForMigration",
+            this.limitCpuForMigration == null ? null : this.limitCpuForMigration.toString());
+        jsonWriter.writeStringField("dynamicMemoryEnabled",
+            this.dynamicMemoryEnabled == null ? null : this.dynamicMemoryEnabled.toString());
+        jsonWriter.writeNumberField("dynamicMemoryMaxMB", this.dynamicMemoryMaxMB);
+        jsonWriter.writeNumberField("dynamicMemoryMinMB", this.dynamicMemoryMinMB);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of HardwareProfileUpdate from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of HardwareProfileUpdate if the JsonReader was pointing to an instance of it, or null if it
+     * was pointing to JSON null.
+     * @throws IOException If an error occurs while reading the HardwareProfileUpdate.
+     */
+    public static HardwareProfileUpdate fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            HardwareProfileUpdate deserializedHardwareProfileUpdate = new HardwareProfileUpdate();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("memoryMB".equals(fieldName)) {
+                    deserializedHardwareProfileUpdate.memoryMB = reader.getNullable(JsonReader::getInt);
+                } else if ("cpuCount".equals(fieldName)) {
+                    deserializedHardwareProfileUpdate.cpuCount = reader.getNullable(JsonReader::getInt);
+                } else if ("limitCpuForMigration".equals(fieldName)) {
+                    deserializedHardwareProfileUpdate.limitCpuForMigration
+                        = LimitCpuForMigration.fromString(reader.getString());
+                } else if ("dynamicMemoryEnabled".equals(fieldName)) {
+                    deserializedHardwareProfileUpdate.dynamicMemoryEnabled
+                        = DynamicMemoryEnabled.fromString(reader.getString());
+                } else if ("dynamicMemoryMaxMB".equals(fieldName)) {
+                    deserializedHardwareProfileUpdate.dynamicMemoryMaxMB = reader.getNullable(JsonReader::getInt);
+                } else if ("dynamicMemoryMinMB".equals(fieldName)) {
+                    deserializedHardwareProfileUpdate.dynamicMemoryMinMB = reader.getNullable(JsonReader::getInt);
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedHardwareProfileUpdate;
+        });
     }
 }